genai/nl-query/requirements.txt
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
duckdb==1.3.2
google-cloud-secret-manager>=2.16.0
langgraph>=1.0.2
langchain>=1.0.3
langchain-google-vertexai>=2.1.2
langsmith>=0.1.0
streamlit-mermaid>=0.0.1

